# BasicApi.IntegrationTests/Snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "AttachmentDto": {
        "additionalProperties": false,
        "description": "A file as clients see it, in messages and on avatars. The bytes are fetched by link.",
        "properties": {
          "durationMs": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "fileName": {
            "nullable": true,
            "type": "string"
          },
          "hasThumbnail": {
            "description": "There is a preview (photo; video when the client sent a frame).",
            "type": "boolean"
          },
          "height": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "kind": {
            "description": "`photo`, `video`, `file` or `voice`.",
            "nullable": true,
            "type": "string"
          },
          "mimeType": {
            "description": "The type the server found in the file; `application/octet-stream` — unknown.",
            "nullable": true,
            "type": "string"
          },
          "size": {
            "format": "int64",
            "type": "integer"
          },
          "state": {
            "description": "`stored`; `expired` — the original was removed by the server's retention policy,\n            only the preview is left.",
            "nullable": true,
            "type": "string"
          },
          "waveform": {
            "items": {
              "format": "int32",
              "type": "integer"
            },
            "nullable": true,
            "type": "array"
          },
          "width": {
            "description": "Photo: as shown (camera rotation applied); video: as the client said.",
            "format": "int32",
            "nullable": true,
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "ChatListItemDto": {
        "additionalProperties": false,
        "properties": {
          "archived": {
            "description": "In the archive: not in the main list. A new message brings it back unless muted.",
            "type": "boolean"
          },
          "avatarId": {
            "description": "The photo to show for the chat: the group's, or the companion's in a private chat; null — none.\nFetched by `POST /api/media/links`.",
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "chatId": {
            "format": "uuid",
            "type": "string"
          },
          "companionId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "companionName": {
            "nullable": true,
            "type": "string"
          },
          "companionUsername": {
            "description": "Companion username (private chats only, null for groups).",
            "nullable": true,
            "type": "string"
          },
          "draft": {
            "$ref": "#/components/schemas/DraftDto"
          },
          "isMuted": {
            "description": "Muted now: no notifications; the unread counter works as usual.",
            "type": "boolean"
          },
          "lastActivityAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastMessage": {
            "$ref": "#/components/schemas/MessageDto"
          },
          "lastReadSeq": {
            "description": "Seq of the last message the user has read.",
            "format": "int64",
            "type": "integer"
          },
          "markedUnread": {
            "description": "The user marked the chat as unread; reading it clears the mark.",
            "type": "boolean"
          },
          "mutedUntil": {
            "description": "Muted until then; null — not muted, or muted for good (BasicApi.Models.Dto.Chat.ChatListItemDto.IsMuted).",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "outboxDeliveredSeq": {
            "description": "The user's own messages up to this seq have reached another member's device; 0 — none.\nGrows with `MessagesDelivered` and `MessagesRead`.",
            "format": "int64",
            "type": "integer"
          },
          "outboxReadSeq": {
            "description": "The user's own messages up to this seq have been read by another member; 0 — none.\nGrows with `MessagesRead`.",
            "format": "int64",
            "type": "integer"
          },
          "pinnedPosition": {
            "description": "Pinned to the top of the list: 1 — the topmost; null — not pinned.",
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "title": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string"
          },
          "unreadCount": {
            "format": "int32",
            "type": "integer"
          },
          "unreadMentionCount": {
            "description": "Unread messages that mention the user.",
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "DraftDto": {
        "additionalProperties": false,
        "description": "An unsent message, the same on all the user's devices.",
        "properties": {
          "entities": {
            "items": {
              "$ref": "#/components/schemas/MessageEntityDto"
            },
            "nullable": true,
            "type": "array"
          },
          "replyToMessageId": {
            "description": "The message the draft answers; null — not a reply.",
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "text": {
            "description": "The text as typed, not trimmed.",
            "nullable": true,
            "type": "string"
          },
          "updatedAt": {
            "description": "When the draft was last saved (server time).",
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "FolderDto": {
        "additionalProperties": false,
        "description": "A folder of chats.",
        "properties": {
          "chatIds": {
            "description": "Chats listed explicitly (the pinned ones among them), archived ones too.",
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "includeGroups": {
            "description": "All groups, outside the archive.",
            "type": "boolean"
          },
          "includePrivate": {
            "description": "All private chats, outside the archive.",
            "type": "boolean"
          },
          "onlyUnread": {
            "description": "Only chats with something unread (the pinned ones always).",
            "type": "boolean"
          },
          "pinnedChatIds": {
            "description": "Chats pinned inside the folder, top first.",
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "title": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "MessageActionDto": {
        "additionalProperties": false,
        "description": "What a system message records.",
        "properties": {
          "title": {
            "description": "The group's title — for `group_created` and `title_changed`.",
            "nullable": true,
            "type": "string"
          },
          "type": {
            "description": "`group_created`, `title_changed`, `members_added`, `member_removed`, `member_left`, `photo_changed`, `photo_removed`.",
            "nullable": true,
            "type": "string"
          },
          "userIds": {
            "description": "Whom it concerns: the added or removed members, the one who left.",
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          }
        },
        "type": "object"
      },
      "MessageDto": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "$ref": "#/components/schemas/MessageActionDto"
          },
          "attachments": {
            "description": "The files, in album order; empty — none. The bytes are fetched by `POST /api/media/links`.",
            "items": {
              "$ref": "#/components/schemas/AttachmentDto"
            },
            "nullable": true,
            "type": "array"
          },
          "chatId": {
            "format": "uuid",
            "type": "string"
          },
          "clientMessageId": {
            "description": "Id supplied by the sender when sending via REST; null — not supplied.",
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "editedAt": {
            "description": "When the text was last edited; null — never edited.",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "entities": {
            "description": "Formatting and mentions over BasicApi.Models.Dto.Message.MessageDto.Text; empty — plain text.",
            "items": {
              "$ref": "#/components/schemas/MessageEntityDto"
            },
            "nullable": true,
            "type": "array"
          },
          "forwardFrom": {
            "$ref": "#/components/schemas/MessageForwardDto"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "isRead": {
            "description": "Someone else's message — the caller has read it; one's own — another member has read it\n(BasicApi.Models.Dto.Message.MessageDto.Status is `read`). Filled in history and search; false in events.",
            "type": "boolean"
          },
          "myReaction": {
            "description": "The caller's own reaction. Filled in history and search; in events it is null —\ntrack your own from `ReactionsChanged`.",
            "nullable": true,
            "type": "string"
          },
          "reactions": {
            "description": "Reactions, most popular first; empty — none.",
            "items": {
              "$ref": "#/components/schemas/ReactionCountDto"
            },
            "nullable": true,
            "type": "array"
          },
          "replyTo": {
            "$ref": "#/components/schemas/MessageReplyDto"
          },
          "senderId": {
            "format": "uuid",
            "type": "string"
          },
          "senderName": {
            "nullable": true,
            "type": "string"
          },
          "seq": {
            "description": "Number of the message in the chat: 1, 2, 3, … Messages are ordered by it.",
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "description": "For the caller's own messages: `sent`, `delivered` (a device of another member\nreceived it) or `read` (another member read it). Null for others' messages, in a chat\nwith oneself and in events. Filled in history and search.",
            "nullable": true,
            "type": "string"
          },
          "text": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "description": "Kind of message: `text`, `media` (files in BasicApi.Models.Dto.Message.MessageDto.Attachments, the text is the\ncaption and may be empty) or `system` (see BasicApi.Models.Dto.Message.MessageDto.Action); more kinds will come —\nshow an unknown one as a stub.",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "MessageEntityDto": {
        "additionalProperties": false,
        "description": "A piece of formatting over the message text. BasicApi.Models.Dto.Message.MessageEntityDto.Offset and BasicApi.Models.Dto.Message.MessageEntityDto.Length\nare in UTF-16 code units — the same as string indices in JavaScript, Java, C# and Swift's\n`utf16` view. A record: two lists of entities compare by value.",
        "properties": {
          "language": {
            "description": "`pre`: language for highlighting, optional.",
            "nullable": true,
            "type": "string"
          },
          "length": {
            "format": "int32",
            "type": "integer"
          },
          "offset": {
            "format": "int32",
            "type": "integer"
          },
          "type": {
            "description": "`bold`, `italic`, `underline`, `strikethrough`, `spoiler`,\n            `code`, `pre` (a code block, with BasicApi.Models.Dto.Message.MessageEntityDto.Language), `link`\n            (with BasicApi.Models.Dto.Message.MessageEntityDto.Url), `mention` (with BasicApi.Models.Dto.Message.MessageEntityDto.UserId).",
            "nullable": true,
            "type": "string"
          },
          "url": {
            "description": "`link`: http, https or mailto.",
            "nullable": true,
            "type": "string"
          },
          "userId": {
            "description": "`mention`: the mentioned member of the chat.",
            "format": "uuid",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "MessageForwardDto": {
        "additionalProperties": false,
        "properties": {
          "senderId": {
            "format": "uuid",
            "type": "string"
          },
          "senderName": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "MessageReplyDto": {
        "additionalProperties": false,
        "description": "Preview of the answered message.",
        "properties": {
          "deleted": {
            "description": "The answered message was deleted for everyone.",
            "type": "boolean"
          },
          "messageId": {
            "format": "uuid",
            "type": "string"
          },
          "senderId": {
            "format": "uuid",
            "type": "string"
          },
          "senderName": {
            "nullable": true,
            "type": "string"
          },
          "text": {
            "description": "The start of the text; empty when the message was deleted.",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "OwnProfileResponseDto": {
        "additionalProperties": false,
        "description": "The caller's own profile: `AuthResponseDto` minus the tokens. Unlike `UserProfileResponseDto`\nit includes the email, which is private to the account owner.",
        "properties": {
          "avatarId": {
            "description": "The photo; null — none. Fetched by `POST /api/media/links`.",
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "displayName": {
            "nullable": true,
            "type": "string"
          },
          "email": {
            "nullable": true,
            "type": "string"
          },
          "userId": {
            "format": "uuid",
            "type": "string"
          },
          "username": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "PrivacySettingsDto": {
        "additionalProperties": false,
        "description": "Privacy settings: `everybody`, `contacts` (those who share a chat with the user) or\n`nobody`. In a request a null field is left as it is.",
        "properties": {
          "groupAdd": {
            "description": "Who may add the user to groups.",
            "nullable": true,
            "type": "string"
          },
          "lastSeen": {
            "description": "Who sees online and last seen. Hiding one's own hides the others' too.",
            "nullable": true,
            "type": "string"
          },
          "messages": {
            "description": "Who may start a private chat; an existing one keeps working.",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "ProblemDetails": {
        "additionalProperties": {},
        "properties": {
          "detail": {
            "nullable": true,
            "type": "string"
          },
          "instance": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "title": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ReactionCountDto": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "format": "int32",
            "type": "integer"
          },
          "emoji": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "SyncAckDto": {
        "additionalProperties": false,
        "properties": {
          "pts": {
            "description": "pts of the last change the device received and processed.",
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "SyncDifferenceDto": {
        "additionalProperties": false,
        "description": "Changes after the pts known to the client.",
        "properties": {
          "hasMore": {
            "description": "There are more changes — repeat the request with the new `since`.",
            "type": "boolean"
          },
          "pts": {
            "description": "The pts up to which the client is now up to date: pass it to the next request as `since`.",
            "format": "int64",
            "type": "integer"
          },
          "snapshotRequired": {
            "description": "No diff available: part of the changes has already been removed from the journal (the client did not\nsign in within the retention period) or `since` is not from this journal. Take a new snapshot via\n`GET /api/sync/state`.",
            "type": "boolean"
          },
          "updates": {
            "description": "Changes in pts order.",
            "items": {
              "$ref": "#/components/schemas/SyncUpdateDto"
            },
            "nullable": true,
            "type": "array"
          }
        },
        "type": "object"
      },
      "SyncStateDto": {
        "additionalProperties": false,
        "description": "Snapshot: the user's state to start from, the same on every device; chat history comes page by page.",
        "properties": {
          "blockedUserIds": {
            "description": "Whom the user has blocked; profiles — `GET /api/users/me/blocked`.",
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "chats": {
            "description": "All the user's chats, archived included — like `GET /api/chats`: counters, pins,\narchive, mute, drafts.",
            "items": {
              "$ref": "#/components/schemas/ChatListItemDto"
            },
            "nullable": true,
            "type": "array"
          },
          "folders": {
            "description": "The user's folders, in order — like `GET /api/folders`.",
            "items": {
              "$ref": "#/components/schemas/FolderDto"
            },
            "nullable": true,
            "type": "array"
          },
          "me": {
            "$ref": "#/components/schemas/OwnProfileResponseDto"
          },
          "privacy": {
            "$ref": "#/components/schemas/PrivacySettingsDto"
          },
          "pts": {
            "description": "Number of the last change included in the snapshot. From it — `GET /api/sync?since=`.",
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "SyncUpdateDto": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "payload": {
            "description": "The same object that arrives in the hub event with this name."
          },
          "pts": {
            "format": "int64",
            "type": "integer"
          },
          "type": {
            "description": "Hub event name: `MessageCreated`, `MessageUpdated`, `MessageDeleted`, `ReactionsChanged`, `MessagesDelivered`, `MessagesRead`, `ReadStateChanged`, `DraftUpdated`, `ChatCreated`, `ChatUpdated`,\n            `ChatDeleted`, `MemberAdded`, `MemberRemoved`, `MemberUpdated`, `UserUpdated`, `PrivacyUpdated`,\n            `BlockListChanged`, `PinnedChatsChanged`, `ChatStateChanged`,\n            `FoldersChanged`.",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "bearerFormat": "JWT",
        "description": "Enter your JWT token. Example: 'eyJhbGci...'",
        "scheme": "Bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "support@myapi.com",
      "name": "API Support Team",
      "url": "https://myapi.com"
    },
    "description": "Real-time Chat API with JWT authentication",
    "title": "Chat API",
    "version": "v1"
  },
  "openapi": "3.0.4",
  "paths": {
...
    "/api/sync": {
      "get": {
        "description": "After a reconnect: apply `updates` in order (each has the same payload as the\nhub event of the same name), store the returned `pts`, repeat while\n`hasMore`. An update may also have arrived over the hub — recognize it by\nthe message or chat id.\n            \n`snapshotRequired: true` — the journal no longer reaches back to\n`since` (kept for 30 days): take `GET /api/sync/state` instead.",
        "parameters": [
          {
            "description": "The last pts the client knows (0 — nothing yet).",
            "in": "query",
            "name": "since",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "description": "Updates per page (default 100, max 500).",
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 100,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncDifferenceDto"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "summary": "Updates after `since`, in order.",
        "tags": [
          "Sync"
        ]
      }
    },
    "/api/sync/ack": {
      "post": {
        "description": "The device is the sign-in behind the access token. The acknowledged pts only moves\nforward. New messages the journal carried up to this pts become `delivered`: their\nauthors get `MessagesDelivered` if no other member had received them before.",
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SyncAckDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncAckDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncAckDto"
              }
            }
          },
          "description": "The last received pts."
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "summary": "Confirm that this device has received updates up to `pts`.",
        "tags": [
          "Sync"
        ]
      }
    },
    "/api/sync/state": {
      "get": {
        "description": "Call on first start and whenever `GET /api/sync` answers\n`snapshotRequired: true`. Keep `pts`: it is the `since` of the next\n`GET /api/sync`.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStateDto"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "summary": "Snapshot to start from: the chat list and the current `pts`.",
        "tags": [
          "Sync"
        ]
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
...
}